# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API_Proj",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Offices/Get/{id}": {
      "get": {
        "tags": [
          "Offices"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OfficeDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfficeDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfficeDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/Offices/Update": {
      "put": {
        "tags": [
          "Offices"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfficeDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OfficeDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OfficeDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Offices/Update/{officeID}/UnassignRegion": {
      "put": {
        "tags": [
          "Offices"
        ],
        "parameters": [
          {
            "name": "officeID",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Offices/Create": {
      "post": {
        "tags": [
          "Offices"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfficeForCreationDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OfficeForCreationDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OfficeForCreationDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OfficeDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfficeDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfficeDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/Offices/Delete/{id}": {
      "delete": {
        "tags": [
          "Offices"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Regions/Get/{id}": {
      "get": {
        "tags": [
          "Regions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RegionDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/Regions/Update": {
      "put": {
        "tags": [
          "Regions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegionDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegionDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegionDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Regions/Create": {
      "post": {
        "tags": [
          "Regions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegionForCreationDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegionForCreationDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegionForCreationDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RegionDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/Regions/Delete/{id}": {
      "delete": {
        "tags": [
          "Regions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "OfficeDTO": {
        "required": [
          "officeID"
        ],
        "type": "object",
        "properties": {
          "officeID": {
            "type": "integer",
            "format": "int32"
          },
          "officeName": {
            "type": "string",
            "nullable": true
          },
          "regionID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "employeesIDs": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OfficeForCreationDTO": {
        "required": [
          "officeName"
        ],
        "type": "object",
        "properties": {
          "officeName": {
            "minLength": 1,
            "type": "string"
          },
          "regionID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "employeesIDs": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RegionDTO": {
        "required": [
          "regionID"
        ],
        "type": "object",
        "properties": {
          "regionID": {
            "type": "integer",
            "format": "int32"
          },
          "regionName": {
            "type": "string",
            "nullable": true
          },
          "officesIDs": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RegionForCreationDTO": {
        "required": [
          "regionName"
        ],
        "type": "object",
        "properties": {
          "regionName": {
            "minLength": 1,
            "type": "string"
          },
          "officesIDs": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}